Recojo Cambio Cantidad de camión a decimal

+++ AcopioAPIs/DTOs/Recojo/RecojoDto.cs
@@ -5,7 +5,7 @@
         public int RecojoId { get; set; }
         public DateOnly RecojoFechaInicio { get; set; }
         public DateOnly RecojoFechaFin { get; set; }
-        public int RecojoCamionesCantidad { get; set; }
+        public decimal RecojoCamionesCantidad { get; set; }
         public decimal RecojoCamionesPrecio { get; set; }
         public int RecojoDiasCantidad { get; set; }
         public decimal RecojoDiasPrecio { get; set; }
