+++ AcopioAPIs/Models/Recojo.cs
@@ -13,7 +13,7 @@ public partial class Recojo
 
     public string? RecojoCampo { get; set; }
 
-    public int RecojoCamionesCantidad { get; set; }
+    public decimal RecojoCamionesCantidad { get; set; }
 
     public decimal RecojoCamionesPrecio { get; set; }
